add link to templating to change display options
The time and date sensor is useful, but the display options are limited unless you understand to use templates. But there is no reference to templates in this page. I propose we add at the very least a link to the templating documentation, and a small snippet of code detailing how to do it.